#  one pole filter optimizations