--- a/Flaky.Sources/Sources/Effects/Filter/OnePoleFilter.cs
+++ b/Flaky.Sources/Sources/Effects/Filter/OnePoleFilter.cs
@@ -9,6 +9,7 @@ namespace Flaky
 		private Source cutoff;
 		private State state;
 		private const int oversampling = 4;
+		private bool isHighPass;
 
 		private class State
 		{
@@ -17,15 +18,17 @@ namespace Flaky
 			public Vector2 latestInputSample;
 		}
 
-		internal OnePoleFilter(Source source, Source cutoff, string id) : base(id)
+		internal OnePoleFilter(Source source, Source cutoff, bool isHighPass, string id) : base(id)
 		{
 			this.source = source;
 			this.cutoff = cutoff;
+			this.isHighPass = isHighPass;
 		}
 
-		internal OnePoleFilter(Source cutoff, string id) : base(id)
+		internal OnePoleFilter(Source cutoff, bool isHighPass, string id) : base(id)
 		{
 			this.cutoff = cutoff;
+			this.isHighPass = isHighPass;
 		}
 
 		protected override void Initialize(IContext context)
@@ -42,39 +45,45 @@ namespace Flaky
 		protected override Vector2 NextSample(IContext context)
 		{
 			var sample = source.Play(context);
-			var cutoffValue = cutoff.Play(context).X;
+			var cutoffValue = cutoff.Play(context).X * 0.25f;
 
 			if (cutoffValue < 0)
 				cutoffValue = 0;
 
-			if (cutoffValue > 1)
-				cutoffValue = 1;
+			if (cutoffValue > 0.25f)
+				cutoffValue = 0.25f;
 
 			var hp = new Vector2(0, 0);
 			var integrator = state.integratorState;
 			var lp = state.integratorState;
-			var s = sample;
 			var latestInput = state.latestInputSample;
 
 			const float oversamplingD = 1 / (float)oversampling;
 
 			for (int i = 1; i <= oversampling; i++)
 			{
-				hp = (s * i + latestInput * (oversampling - i)) * oversamplingD - lp;
+				// lp = (s * i + latestInput * (oversampling - i)) - lp;
+				if (i == 1)
+					hp = sample + latestInput + latestInput + latestInput - lp;
+				else if (i == 2)
+					hp = sample + sample + latestInput + latestInput - lp;
+				else if (i == 3)
+					hp = sample + sample + sample + latestInput - lp;
+				else if (i == 4)
+					hp = sample + sample + sample + sample - lp;
 
-				var input = hp * (cutoffValue * 0.25f);
-				var output = input + integrator;
-
-				integrator = input + output;
-
-				lp = output;
+				lp = hp * cutoffValue + integrator;
+				integrator = hp * cutoffValue + lp;
 			}
 
 			state.latestInputSample = sample;
 			state.lp = lp;
 			state.integratorState = integrator;
 
-			return GetResult(state.lp, hp);
+			if(isHighPass)
+				return hp * oversamplingD;
+			else
+				return lp * oversamplingD;
 		}
 
 		void IPipingSource<Source>.SetMainSource(Source mainSource)