--- a/Flaky.Sources/Sources/Effects/Filter/OnePoleHPFilter.cs
+++ b/Flaky.Sources/Sources/Effects/Filter/OnePoleHPFilter.cs
@@ -4,8 +4,9 @@ namespace Flaky
 {
 	public class OnePoleHPFilter : OnePoleFilter
 	{
-		internal OnePoleHPFilter(Source source, Source cutoff, string id) : base(source, cutoff, id) { }
-		internal OnePoleHPFilter(Source cutoff, string id) : base(cutoff, id) { }
+		internal OnePoleHPFilter(Source source, Source cutoff, string id) 
+			: base(source, cutoff, true, id) { }
+		internal OnePoleHPFilter(Source cutoff, string id) : base(cutoff, true, id) { }
 
 		protected override Vector2 GetResult(Vector2 lp, Vector2 hp)
 		{