--- a/Flaky.Sources/Sources/Effects/Filter/OnePoleHPFilter.cs
+++ b/Flaky.Sources/Sources/Effects/Filter/OnePoleHPFilter.cs
@@ -7,10 +7,5 @@ namespace Flaky
 		internal OnePoleHPFilter(Source source, Source cutoff, string id) 
 			: base(source, cutoff, true, id) { }
 		internal OnePoleHPFilter(Source cutoff, string id) : base(cutoff, true, id) { }
-
-		protected override Vector2 GetResult(Vector2 lp, Vector2 hp)
-		{
-			return hp;
-		}
 	}
 }