--- a/Flaky.Sources/Sources/Effects/Filter/OnePoleLPFilter.cs
+++ b/Flaky.Sources/Sources/Effects/Filter/OnePoleLPFilter.cs
@@ -4,9 +4,11 @@ namespace Flaky
 {
 	public class OnePoleLPFilter : OnePoleFilter
 	{
-		internal OnePoleLPFilter(Source source, Source cutoff, string id) : base(source, cutoff, id) { }
+		internal OnePoleLPFilter(Source source, Source cutoff, string id) 
+			: base(source, cutoff, true, id) { }
 
-		internal OnePoleLPFilter(Source cutoff, string id) : base(cutoff, id) { }
+		internal OnePoleLPFilter(Source cutoff, string id) 
+			: base(cutoff, true, id) { }
 
 		protected override Vector2 GetResult(Vector2 lp, Vector2 hp)
 		{