--- a/Flaky.Sources/Sources/Effects/Filter/OnePoleLPFilter.cs
+++ b/Flaky.Sources/Sources/Effects/Filter/OnePoleLPFilter.cs
@@ -5,14 +5,9 @@ namespace Flaky
 	public class OnePoleLPFilter : OnePoleFilter
 	{
 		internal OnePoleLPFilter(Source source, Source cutoff, string id) 
-			: base(source, cutoff, true, id) { }
+			: base(source, cutoff, false, id) { }
 
 		internal OnePoleLPFilter(Source cutoff, string id) 
-			: base(cutoff, true, id) { }
-
-		protected override Vector2 GetResult(Vector2 lp, Vector2 hp)
-		{
-			return lp;
-		}
+			: base(cutoff, false, id) { }
 	}
 }