filter workload before:

```
             Method |     Mean |     Error |    StdDev |
------------------- |---------:|----------:|----------:|
 FilterWorkloadTest | 33.93 ms | 0.0107 ms | 0.0089 ms |
```

filter workload after:

```
              Method |     Mean |     Error |    StdDev |
------------------- |---------:|----------:|----------:|
 FilterWorkloadTest | 29.76 ms | 0.0273 ms | 0.0242 ms |
```

standard workload before:

```
StandardWorkload.StandardWorkloadTest: DefaultJob
Runtime = .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3260.0; GC = Concurrent Workstation
Mean = 250.3600 ms, StdErr = 4.1073 ms (1.64%); N = 100, StdDev = 41.0733 ms
Min = 206.9401 ms, Q1 = 219.0288 ms, Median = 230.4844 ms, Q3 = 295.6021 ms, Max = 334.6499 ms
IQR = 76.5733 ms, LowerFence = 104.1688 ms, UpperFence = 410.4621 ms
ConfidenceInterval = [236.4298 ms; 264.2901 ms] (CI 99.9%), Margin = 13.9301 ms (5.56% of Mean)
Skewness = 0.72, Kurtosis = 2.07, MValue = 3.83
-------------------- Histogram --------------------
[205.804 ms ; 221.290 ms) | @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
[221.290 ms ; 239.703 ms) | @@@@@@@@
[239.703 ms ; 255.455 ms) |
[255.455 ms ; 270.941 ms) | @@@@@@@@@@@@@@@@@@
[270.941 ms ; 293.766 ms) |
[293.766 ms ; 309.252 ms) | @@@@@@@@@@@@@@@@@@
[309.252 ms ; 326.212 ms) |
[326.212 ms ; 342.393 ms) | @@@@@@@@
---------------------------------------------------
```

standard workload after:

```
StandardWorkload.StandardWorkloadTest: DefaultJob
Runtime = .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3260.0; GC = Concurrent Workstation
Mean = 196.6964 ms, StdErr = 1.0324 ms (0.52%); N = 25, StdDev = 5.1622 ms
Min = 188.5776 ms, Q1 = 192.4465 ms, Median = 198.2948 ms, Q3 = 201.1200 ms, Max = 204.3966 ms
IQR = 8.6735 ms, LowerFence = 179.4362 ms, UpperFence = 214.1302 ms
ConfidenceInterval = [192.8295 ms; 200.5632 ms] (CI 99.9%), Margin = 3.8669 ms (1.97% of Mean)
Skewness = -0.02, Kurtosis = 1.64, MValue = 2
-------------------- Histogram --------------------
[187.808 ms ; 192.759 ms) | @@@@@@
[192.759 ms ; 200.616 ms) | @@@@@@@@@@@@@
[200.616 ms ; 205.941 ms) | @@@@@@
---------------------------------------------------
```